Create a new module for error-handler (#1469)
