+++ pom.xml
@@ -33,6 +33,7 @@
         <module>elasticjob-api</module>
         <module>elasticjob-infra</module>
         <module>elasticjob-executor</module>
+        <module>elasticjob-error-handler</module>
         <module>elasticjob-lite</module>
         <module>elasticjob-cloud</module>
         <module>elasticjob-distribution</module>
